Make Machine.destroy() block
Block the current coro until the machine is actually removed
from the remote model. This follows the same pattern as other
operations, e.g. Model.deploy().

Also clean up tests, making each functional test run on its
own clean model.
