+++ tests/client/test_client.py
@@ -1,24 +1,22 @@
 import pytest
 
-from juju.client.connection import Connection
 from juju.client import client
 
-from ..base import bootstrapped
+from .. import base
 
 
-@bootstrapped
+@base.bootstrapped
 @pytest.mark.asyncio
 async def test_user_info(event_loop):
-    conn = await Connection.connect_current()
-    controller_conn = await conn.controller()
+    async with base.CleanModel() as model:
+        controller_conn = await model.connection.controller()
 
-    um = client.UserManagerFacade()
-    um.connect(controller_conn)
-    result = await um.UserInfo(
-        [client.Entity('user-admin')], True)
-    await conn.close()
-    await controller_conn.close()
+        um = client.UserManagerFacade()
+        um.connect(controller_conn)
+        result = await um.UserInfo(
+            [client.Entity('user-admin')], True)
+        await controller_conn.close()
 
-    assert isinstance(result, client.UserInfoResults)
-    for r in result.results:
-        assert isinstance(r, client.UserInfoResult)
+        assert isinstance(result, client.UserInfoResults)
+        for r in result.results:
+            assert isinstance(r, client.UserInfoResult)
